requirements.txt
telethon
telethon[cryptg] 
pyaes
tgcrypto
pillow
cryptg
aiohttp
hachoir
aioredis
redis
pyyaml
loguru
envparse
motor
aiosqlite
# sqlite3
mysql-connector-python
pandas
sqlalchemy
babel
telegraph
markdown
carbonnow
bs4
selenium
colour
colormap
easydev